Fix macos builds

+++ src/libhictk/balancing/include/hictk/balancing/ice.hpp
@@ -139,18 +139,18 @@ class ICE {
   static void min_count_filtering(nonstd::span<double> biases, std::size_t min_count,
                                   nonstd::span<const double> marg);
 
-  static void mad_max_filtering(nonstd::span<const std::size_t> chrom_offsets,
+  static void mad_max_filtering(nonstd::span<const std::uint64_t> chrom_offsets,
                                 nonstd::span<double> biases, nonstd::span<double> marg,
                                 double mad_max);
 
   template <typename MatrixT>
   static void initialize_biases(const MatrixT& matrix, nonstd::span<double> biases,
-                                nonstd::span<const std::size_t> chrom_bin_offsets,
+                                nonstd::span<const std::uint64_t> chrom_bin_offsets,
                                 std::size_t min_nnz, std::size_t min_count, double mad_max,
                                 BS::thread_pool* tpool);
 
   [[nodiscard]] static std::vector<double> compute_weights_from_chromosome_sizes(
-      const BinTable& bins, nonstd::span<std::size_t> chrom_bin_offsets);
+      const BinTable& bins, nonstd::span<std::uint64_t> chrom_bin_offsets);
 };
 
 }  // namespace hictk::balancing
